+++ src/map.c
@@ -15,31 +15,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void	print_map(t_map *m)
-{
-	int	i;
-	int	s;
-	int	pp;
-
-	printf("Map size: (%d,%d)\n", m->width, m->height);
-	printf("Player Position: (%d,%d)\n",
-		m->player_spawn_x, m->player_spawn_y);
-	i = 0;
-	s = m->width * m->height;
-	pp = m->player_spawn_x + m->width * m->player_spawn_y;
-	while (i < s)
-	{
-		if (i == pp)
-			printf("%c", m->player_facing);
-		else
-			(printf("%c", m->map[i]));
-		++i;
-		if (!(i % m->width) && i < s - 1)
-			printf("\n");
-	}
-	printf("\n");
-}
-
 static int	is_valid_line(char *l, int *found_player)
 {
 	int	i;
@@ -59,14 +34,6 @@ static int	is_valid_line(char *l, int *found_player)
 		++i;
 	}
 	return (i);
-}
-
-void	destroy_map(t_map *m)
-{
-	free(m->map);
-	m->map = 0;
-	m->width = 0;
-	m->height = 0;
 }
 
 // Sets the map size and
@@ -97,49 +64,45 @@ int	set_map_size(t_map *m, t_list *l)
 	return (1);
 }
 
-int	pos_to_index(int width, int x, int y)
+void	set_player(t_map *m, int x, int y, char f)
 {
-	return (y * width + x);
+	m->map[y * m->width + x] = '0';
+	m->player_spawn_x = x;
+	m->player_spawn_y = y;
+	m->player_facing = f;
 }
 
-int	is_map_enclosed(t_map *m)
+void	populate_map(t_map *m, t_list *l)
 {
-	int	x;
-	int	y;
+	char	*line;
+	int		read_line;
+	int		y;
+	int		x;
 
-	y = 0;
-	while (++y < m->height - 1)
+	y = -1;
+	while (++y < m->height)
 	{
-		x = 0;
-		while (++x < m->width - 1)
-			if (m->map[pos_to_index(m->width, x, y)] == '0'
-					&& (m->map[pos_to_index(m->width, x - 1, y)] == ' '
-					|| m->map[pos_to_index(m->width, x - 1, y - 1)] == ' '
-					|| m->map[pos_to_index(m->width, x - 1, y + 1)] == ' '
-					|| m->map[pos_to_index(m->width, x + 1, y - 1)] == ' '
-					|| m->map[pos_to_index(m->width, x + 1, y)] == ' '
-					|| m->map[pos_to_index(m->width, x + 1, y + 1)] == ' '
-					|| m->map[pos_to_index(m->width, x, y + 1)] == ' '
-					|| m->map[pos_to_index(m->width, x, y - 1)] == ' '))
-				return (0);
-		++y;
+		line = (char *)(l->content);
+		read_line = 0;
+		x = -1;
+		while (++x < m->width)
+		{
+			if (line[x] == '\0' || line[x] == '\n')
+				read_line = 1;
+			if (!read_line && is_player_char(line[x]))
+				set_player(m, x, y, line[x]);
+			else if (!read_line)
+				m->map[y * m->width + x] = line[x];
+			else
+				m->map[y * m->width + x] = ' ';
+		}
+		l = l->next;
 	}
-	x = -1;
-	while (++x < m->height)
-		if (m->map[pos_to_index(m->width, 0, x)] == '0'
-			|| m->map[pos_to_index(m->width, m->width - 1, x)] == '0'
-			|| m->map[pos_to_index(m->width, x, 0)] == '0'
-			|| m->map[pos_to_index(m->width, x, m->height - 1)] == '0')
-			return (0);
-	return (1);
 }
 
 t_map	*get_map(t_list *l)
 {
 	t_map	*m;
-	int		x;
-	int		y;
-	int		read_line;
 
 	m = malloc(sizeof(t_map));
 	if (!m || !set_map_size(m, l) || m->width < 3 || m->height < 3)
@@ -151,36 +114,9 @@ t_map	*get_map(t_list *l)
 	if (!m->map)
 	{
 		free(m);
-		return(0);
+		return (0);
 	}
-	y = 0;
-	while (y < m->height)
-	{
-		read_line = 0;
-		x = 0;
-		while (x < m->width)
-		{
-			if (((char *)(l->content))[x] == '\0' || ((char *)(l->content))[x] == '\n')
-				read_line = 1;
-			if (!read_line && (((char *)(l->content))[x] == 'N'
-				|| ((char *)(l->content))[x] == 'S'
-				|| ((char *)(l->content))[x] == 'E'
-				|| ((char *)(l->content))[x] == 'W'))
-			{
-				m->map[y * m->width + x] = '0';
-				m->player_spawn_x = x;
-				m->player_spawn_y = y;
-				m->player_facing = ((char *)(l->content))[x];
-			}
-			else if (!read_line)
-				m->map[y * m->width + x] = ((char *)(l->content))[x];
-			else
-				m->map[y * m->width + x] = ' ';
-			x++;
-		}
-		l = l->next;
-		y++;
-	}
+	populate_map(m, l);
 	if (!is_map_enclosed(m))
 	{
 		destroy_map(m);
